+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 * Go: Add Function Kill ([#3604](https://github.com/valkey-io/valkey-glide/pull/3604))
 * Go: Script eval/load([#3605](https://github.com/valkey-io/valkey-glide/pull/3605))
 * Benchmarks: Fix rust benchmark latencies calculation
+* Go: Add Script Exists ([#3649](https://github.com/valkey-io/valkey-glide/pull/3649))
 
 #### Breaking Changes